+++ zephyr-mcumgr-cli/src/groups/mcuboot.rs
@@ -1,4 +1,7 @@
-use crate::{args::CommonArgs, client::Client, errors::CliError, file_read_write::read_input_file};
+use crate::{
+    args::CommonArgs, client::Client, errors::CliError, file_read_write::read_input_file,
+    formatting::structured_print,
+};
 
 #[derive(Debug, clap::Subcommand)]
 pub enum MCUbootCommand {
@@ -9,12 +12,17 @@ pub enum MCUbootCommand {
     },
 }
 
-pub fn run(_client: &Client, _args: CommonArgs, command: MCUbootCommand) -> Result<(), CliError> {
+pub fn run(_client: &Client, args: CommonArgs, command: MCUbootCommand) -> Result<(), CliError> {
     match command {
         MCUbootCommand::GetImageInfo { file } => {
-            let (data, _source_filename) = read_input_file(&file)?;
+            let (image_data, _source_filename) = read_input_file(&file)?;
+            let image_info =
+                zephyr_mcumgr::mcuboot::get_image_info(std::io::Cursor::new(image_data.as_ref()))?;
 
-            println!("Image size: {}", data.len());
+            structured_print(Some(file), args.json, |s| {
+                s.key_value("version", image_info.version.to_string());
+                s.key_value("hash", hex::encode(image_info.hash));
+            })?;
         }
     }