+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - xxxx-xx-xx
+
+- Add MCUboot firmware image parser
+  - Rust: `mcuboot::get_image_info`
+  - Python: `mcuboot_get_image_info`
+  - CLI: `mcuboot get-image-info`
+
 ## [0.5.1] - 2025-12-07
 
 ### Changes
@@ -177,6 +184,7 @@ Initial release, not feature complete yet.
 
 Primarily to test release workflow.
 
+[0.5.2]: https://github.com/Finomnis/zephyr-mcumgr-client/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/Finomnis/zephyr-mcumgr-client/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/Finomnis/zephyr-mcumgr-client/compare/0.4.2...0.5.0
 [0.4.2]: https://github.com/Finomnis/zephyr-mcumgr-client/compare/0.4.1...0.4.2